requirements.txt
# I'm not mentioning the specific verison of these right now but its best practive to pin versions in production.
fastapi
uvicorn
pydantic
pydantic-settings

portia-sdk-python[all]==0.7.2

sqlalchemy

PyGitHub
requests

gitpython
python-multipart

jinja2

python-dotenv
pyyaml

httpx
aiofiles

pytest
black
pre-commit
